api-key-generator/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
